+++ ml_tools/imageprocessing.py
@@ -2,7 +2,7 @@ import cv2
 from pathlib import Path
 import numpy as np
 from PIL import Image, ImageDraw
-
+import math
 from ml_tools.tools import eucl_distance
 from track.track import TrackChannels
 from scipy import ndimage
@@ -142,3 +142,42 @@ def resize_cv(image, dim, interpolation=cv2.INTER_LINEAR, extra_h=0, extra_v=0):
         dsize=(dim[0] + extra_h, dim[1] + extra_v),
         interpolation=interpolation,
     )
+
+
+def filtered_is_valid(frame, label):
+    filtered = frame.filtered
+    thermal = frame.thermal
+    thermal_deviation = np.amax(thermal) != np.amin(thermal)
+    filtered_deviation = np.amax(filtered) != np.amin(filtered)
+    if not thermal_deviation or not filtered_deviation:
+        return False
+    if label == "false-positive":
+        return True
+
+    area = filtered.shape[0] * filtered.shape[1]
+    percentile = int(100 - 100 * 16.0 / area)
+    threshold = np.percentile(filtered, percentile)
+    threshold = max(0, threshold - 40)
+
+    rows = math.floor(0.1 * filtered.shape[0])
+    columns = math.floor(0.1 * filtered.shape[1])
+    rows = np.clip(rows, 1, 2)
+    columns = np.clip(columns, 1, 2)
+
+    top_left = 1 if np.amax(filtered[0:rows][:, 0:columns]) > threshold else 0
+    top_right = 1 if np.amax(filtered[0:rows][:, -columns - 1 : -1]) > threshold else 0
+    bottom_left = (
+        1 if np.amax(filtered[-rows - 1 : -1][:, 0:columns]) > threshold else 0
+    )
+    bottom_right = (
+        1 if np.amax(filtered[-rows - 1 : -1][:, -columns - 1 : -1]) > threshold else 0
+    )
+    # try and filter out bogus frames where data is on 3 or more corners
+    if (top_right + bottom_left + top_left + bottom_right) >= 3:
+        return False
+
+    num_less = len(filtered[filtered <= threshold])
+
+    if num_less <= area * 0.05 or np.amax(filtered) == np.amin(filtered):
+        return False
+    return True
